requirements.txt
APScheduler==3.10.1
beautifulsoup4==4.12.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==21.2.0
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.20.1
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
qrcode==6.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
youtube-data-api==0.0.20
